requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.52.0
django-simple-chatbot==0.0.9
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
whitenoise==4.1
wrapt==1.11.1
